add products perspective
