+++ codbex-inventory/codbex-inventory.model
@@ -1279,9 +1279,9 @@
                 "name": "Product",
                 "perspectiveHeader": "",
                 "perspectiveIcon": "/services/web/resources/unicons/cube.svg",
-                "perspectiveLabel": "Entities",
+                "perspectiveLabel": "Products",
                 "perspectiveName": "Products",
-                "perspectiveNavId": "",
+                "perspectiveNavId": "products-navigation-group",
                 "perspectiveOrder": "100",
                 "perspectiveRole": "",
                 "projectionReferencedEntity": "Product",
@@ -3861,7 +3861,7 @@
                 "perspectiveIcon": "/services/web/resources/unicons/cube.svg",
                 "perspectiveLabel": "Products",
                 "perspectiveName": "Products",
-                "perspectiveNavId": "",
+                "perspectiveNavId": "products-navigation-group",
                 "perspectiveOrder": "100",
                 "perspectiveRole": "",
                 "projectionReferencedEntity": "ProductPackaging",
@@ -3972,10 +3972,10 @@
                 "menuLabel": "ProductAvailability",
                 "name": "ProductAvailability",
                 "perspectiveHeader": "",
-                "perspectiveIcon": "/services/web/resources/unicons/copy.svg",
-                "perspectiveLabel": "Entities",
-                "perspectiveName": "ProductAvailability",
-                "perspectiveNavId": "",
+                "perspectiveIcon": "/services/web/resources/unicons/cube.svg",
+                "perspectiveLabel": "Products",
+                "perspectiveName": "Products",
+                "perspectiveNavId": "products-navigation-group",
                 "perspectiveOrder": "100",
                 "perspectiveRole": "",
                 "title": "ProductAvailability",
@@ -4049,6 +4049,13 @@
                 "label": "Waste",
                 "icon": "/services/web/resources/unicons/trash.svg",
                 "order": "750"
+            },
+            {
+                "name": "Products",
+                "label": "Products",
+                "navId": "products-navigation-group",
+                "icon": "/services/web/resources/unicons/cube.svg",
+                "order": "100"
             }
         ],
         "navigations": []
