--- codbex-inventory/gen/codbex-inventory/dao/ProductAvailability/ProductAvailability.extensionpoint
@@ -1,4 +1,0 @@
-{
-    "name": "codbex-inventory-ProductAvailability-ProductAvailability",
-    "description": "Extension Point for the codbex-inventory-ProductAvailability-ProductAvailability entity"
-}
