+++ codbex-inventory/gen/codbex-inventory/dao/Products/ProductAvailability.extensionpoint
@@ -1,0 +1,4 @@
+{
+    "name": "codbex-inventory-Products-ProductAvailability",
+    "description": "Extension Point for the codbex-inventory-Products-ProductAvailability entity"
+}
