--- codbex-inventory/gen/codbex-inventory/ui/ProductAvailability/Product/ProductAvailability/dialog-filter/view.extension
@@ -1,5 +1,0 @@
-{
-    "module": "codbex-inventory/gen/codbex-inventory/ui/ProductAvailability/Product/ProductAvailability/dialog-filter/view.js",
-    "extensionPoint": "application-windows",
-    "description": "codbex-inventory - Application Dialog Window"
-}
