--- codbex-inventory/gen/codbex-inventory/ui/ProductAvailability/Product/ProductAvailability/dialog-filter/view.js
@@ -1,14 +1,0 @@
-/*
- * Generated by Eclipse Dirigible based on model and template.
- *
- * Do not modify the content as it may be re-generated again.
- */
-const viewData = {
-    id: 'ProductAvailability-filter',
-    label: 'ProductAvailability Filter',
-    path: '/services/web/codbex-inventory/gen/codbex-inventory/ui/ProductAvailability/Product/ProductAvailability/dialog-filter/index.html',
-    perspectiveName: 'ProductAvailability'
-};
-if (typeof exports !== 'undefined') {
-    exports.getView = () => viewData;
-}
