--- codbex-inventory/gen/codbex-inventory/ui/ProductAvailability/Product/ProductAvailability/view.extension
@@ -1,5 +1,0 @@
-{
-    "module": "codbex-inventory/gen/codbex-inventory/ui/ProductAvailability/Product/ProductAvailability/view.js",
-    "extensionPoint": "application-views",
-    "description": "codbex-inventory - Application View - Details"
-}
