+++ codbex-inventory/gen/codbex-inventory/ui/Products/Product/ProductAvailability/dialog-filter/view.extension
@@ -1,0 +1,5 @@
+{
+    "module": "codbex-inventory/gen/codbex-inventory/ui/Products/Product/ProductAvailability/dialog-filter/view.js",
+    "extensionPoint": "application-windows",
+    "description": "codbex-inventory - Application Dialog Window"
+}
