+++ codbex-inventory/gen/codbex-inventory/ui/Products/Product/ProductAvailability/dialog-window/index.html
