+++ codbex-inventory/gen/codbex-inventory/ui/Products/Product/ProductAvailability/index.html
