+++ codbex-inventory/gen/codbex-inventory/ui/Products/Product/ProductAvailability/view.extension
@@ -1,0 +1,5 @@
+{
+    "module": "codbex-inventory/gen/codbex-inventory/ui/Products/Product/ProductAvailability/view.js",
+    "extensionPoint": "application-views",
+    "description": "codbex-inventory - Application View - Details"
+}
