+++ codbex-inventory/gen/codbex-inventory/ui/Products/Product/ProductAvailability/view.js
@@ -1,0 +1,17 @@
+/*
+ * Generated by Eclipse Dirigible based on model and template.
+ *
+ * Do not modify the content as it may be re-generated again.
+ */
+const viewData = {
+    id: 'ProductAvailability',
+    label: 'ProductAvailability',
+    region: 'bottom',
+    lazyLoad: false,
+	autoFocusTab: true,
+    path: '/services/web/codbex-inventory/gen/codbex-inventory/ui/Products/Product/ProductAvailability/index.html',
+    perspectiveName: 'Products'
+};
+if (typeof exports !== 'undefined') {
+	exports.getView = () => viewData;
+}
